Add storage operations utiltiy calss
